requirements.txt
aiohttp==3.8.4
pyjwt==2.7.0
python-dotenv==1.0.0
